+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # Change Log
 #### 2018-01-06 16:30
 * fix bug: forget to check if any key of globalStateKeys is included in global state
+* feature add: startup allow config moduleSingleClass to control some module to only allow register one react class
 
 #### 2018-01-06 16:00
 * fix bug: dispatch action to other module in a reducer block, ccInstance can not recover its globalState correctly, problem found in extractStateToBeBroadcasted